Refactor notebook translator
