+++ src/co_op_translator/core/llm/jupyter_notebook_translator.py
@@ -14,6 +14,7 @@ from .markdown_translator import MarkdownTranslator
 from co_op_translator.utils.common.metadata_utils import (
     calculate_file_hash,
     calculate_string_hash,
+    add_notebook_metadata,
 )
 
 logger = logging.getLogger(__name__)
@@ -127,6 +128,12 @@ class JupyterNotebookTranslator:
             f"Translated {translated_cells}/{total_markdown_cells} markdown cells "
             f"in {notebook_path.name}"
         )
+
+        # Add coopTranslator metadata to the notebook
+        notebook_path = Path(notebook_path)
+        notebook = add_notebook_metadata(notebook, notebook_path, language_code, self.root_dir)
+        
+        logger.debug(f"Added coopTranslator metadata to {notebook_path.name}")
 
         # Return the modified notebook as JSON string
         return json.dumps(notebook, ensure_ascii=False, indent=1)
